Big kludge to enable php4 (and possibly other apps) LFS support.
apr defines _LARGEFILE64_SOURCE which is enough to do LFS with
stat64(), etc., but not regular syscalls like stat().  By not defining
_LARGEFILE_SOURCE or _FILE_OFFSET_BITS=64, apr leaves the choice to the
apps whether to enable those.  Fine.

The problem is that 'ino_t' appears in apr_file_info.h, and if an app
uses the LFS defines, ino_t suddenly changes to a 64-bit type ... which
of course changes the ABI declared in /usr/include so it no longer
matches the real ABI.

The fix is to hard-code ino_t as 'unsigned long', which I believe is
correct on all Debian platforms.  The ABI is the same as it was before,
and now it won't change when an app uses LFS defines.  I've verified
that this fixes #397402 on php4 4:4.4.4-5.

'debian/ino_t_test.c' verifies that this kludge works on a given
platform.  It started life as an autoconf test, but that requires
regenerating configure, a 1 MB diff (100 kB compressed), which I
thought was a bit much to add to the diff.gz.


git-svn-id: svn+ssh://svn.debian.org/svn/pkg-apache/trunk/apr@282 01b336ce-410b-0410-9a02-a0e7f243c266