Return success if socket is already closed
Prior to this commit, socket_cleanup would return an error if the
socket had already been closed.  With this commit we return success
instead.

This is useful, for example, if a socket is closed manually and we
then try to close it again as part of the cleanup of a pool.

Signed-off-by: Anthony Minessale <anthm@freeswitch.org>
Signed-off-by: Travis Cross <tc@traviscross.com>