Prevent endless loop on pool cleanup
When APR_POOL_DEBUG is set, we check up to three levels deep for a
kind of corruption of the cleanup lists where a node is duplicated
causing an endless loop and abort.

If we hit this class of corruption when not debugging, however, we'd
prefer to just ignore it in preference to looping infinitely or
aborting.

With this change, we check for immediate duplicates and simply stop
processing the cleanup list at that point potentially leaking
resources.

Signed-off-by: Anthony Minessale <anthm@freeswitch.org>
Signed-off-by: Travis Cross <tc@traviscross.com>