Add apr_pool_mutex_set()
This commit allows adding a mutex to a pool to make low level
allocations thread-safe.

FreeSWITCH users running large production systems have been using this
technique for over five years without incident.

Since the fork, we've added in tree debugging code to check for
contentious use of a pool.  Here we're able to use those hooks to
reduce the size of the present commit.

Signed-off-by: Anthony Minessale <anthm@freeswitch.org>
Signed-off-by: Travis Cross <tc@traviscross.com>